move logs to kubectl/util
