+++ cmd/gke-certificates-controller/BUILD
@@ -14,7 +14,7 @@ go_library(
     tags = ["automanaged"],
     deps = [
         "//cmd/gke-certificates-controller/app:go_default_library",
-        "//pkg/util/logs:go_default_library",
+        "//pkg/kubectl/util/logs:go_default_library",
         "//pkg/version/verflag:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/util/flag:go_default_library",
