+++ cmd/kubectl/app/BUILD
@@ -16,7 +16,7 @@ go_library(
         "//pkg/client/metrics/prometheus:go_default_library",
         "//pkg/kubectl/cmd:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
-        "//pkg/util/logs:go_default_library",
+        "//pkg/kubectl/util/logs:go_default_library",
         "//pkg/version/prometheus:go_default_library",
         "//vendor/k8s.io/client-go/plugin/pkg/client/auth:go_default_library",
     ],
