+++ federation/cmd/kubefed/app/BUILD
@@ -15,7 +15,7 @@ go_library(
         "//federation/pkg/kubefed:go_default_library",
         "//pkg/client/metrics/prometheus:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
-        "//pkg/util/logs:go_default_library",
+        "//pkg/kubectl/util/logs:go_default_library",
         "//pkg/version:go_default_library",
         "//pkg/version/prometheus:go_default_library",
     ],
