+++ pkg/util/BUILD
@@ -33,7 +33,6 @@ filegroup(
         "//pkg/util/keymutex:all-srcs",
         "//pkg/util/labels:all-srcs",
         "//pkg/util/limitwriter:all-srcs",
-        "//pkg/util/logs:all-srcs",
         "//pkg/util/maps:all-srcs",
         "//pkg/util/metrics:all-srcs",
         "//pkg/util/mount:all-srcs",
